moving more of fusion from vm
