+++ crates/runmat-ignition/src/vm.rs
@@ -3,15 +3,13 @@ use crate::instr::Instr;
 #[cfg(test)]
 use crate::instr::EndExpr;
 #[cfg(feature = "native-accel")]
-use runmat_accelerate::fusion_exec::execute_reduction;
 #[cfg(feature = "native-accel")]
 use runmat_accelerate::{
     activate_fusion_plan, deactivate_fusion_plan, set_current_pc,
 };
 #[cfg(feature = "native-accel")]
 use runmat_accelerate::{
-    active_group_plan_clone, value_is_all_keyword, ReductionAxes, ShapeInfo, ValueOrigin,
-    VarKind,
+    active_group_plan_clone,
 };
 use runmat_builtins::Value;
 use runmat_runtime::{
@@ -849,555 +847,15 @@ async fn try_execute_fusion_group(
     if plan.group.kind.is_elementwise() {
         accel_fusion::execute_fusion_elementwise(request, stack_guard, vars, context)
     } else if plan.group.kind.is_reduction() {
-        // Determine reduction axis or 'all'. Prefer the builtin reduction op's dim argument (inputs[1]).
-        // MATLAB dim is 1-based: dim=1 reduces rows (axis 0), dim=2 reduces cols (axis 1), 'all' reduces all elements.
-        let mut axis = 0usize;
-        let mut axis_explicit = false;
-        let mut reduce_all = matches!(plan.reduction_axes, Some(ReductionAxes::All));
-        if let Some(ReductionAxes::Explicit(dims)) = &plan.reduction_axes {
-            if let Some(first) = dims.first().copied() {
-                axis = first.saturating_sub(1);
-                axis_explicit = true;
-            }
-        }
-        // Debug: show input origins for reduction
-        if log::log_enabled!(log::Level::Debug) {
-            let meta: Vec<String> = plan
-                .inputs
-                .iter()
-                .map(|vid| {
-                    if let Some(info) = graph.value(*vid) {
-                        format!(
-                            "vid={} origin={:?} shape={:?}",
-                            vid, info.origin, info.shape
-                        )
-                    } else {
-                        format!("vid={} origin=<missing>", vid)
-                    }
-                })
-                .collect();
-            log::debug!("reduction gather meta: [{}]", meta.join(", "));
-        }
-        // Detect 'all' in constants or const_values
-        let has_all = reduce_all
-            || plan.constants.values().any(value_is_all_keyword)
-            || plan.const_values.values().any(value_is_all_keyword);
-        if has_all {
-            reduce_all = true;
-        }
-        if reduce_all && interp_engine::fusion_debug_enabled() {
-            log::debug!(
-                "fusion reduction (all) meta: data_vid={:?} inputs={:?} stack_pattern={:?}",
-                plan.reduction_data,
-                plan.inputs,
-                plan.stack_pattern
-            );
-        }
-        if !reduce_all {
-            for node_id in &plan.group.nodes {
-                if let Some(node) = graph.node(*node_id) {
-                    if let runmat_accelerate::graph::AccelNodeLabel::Builtin { name } = &node.label
-                    {
-                        if name.eq_ignore_ascii_case("mean") {
-                            for input_vid in &node.inputs {
-                                if let Some(info) = graph.value(*input_vid) {
-                                    if let Some(constant) = &info.constant {
-                                        if value_is_all_keyword(constant) {
-                                            reduce_all = true;
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-                if reduce_all {
-                    break;
-                }
-            }
-        }
-        // Prefer plan.reduction_dim if available
-        if !reduce_all {
-            if let Some(dim_vid) = plan.reduction_dim {
-                if let Some(cv) = plan.const_values.get(&dim_vid) {
-                    axis = match cv {
-                        Value::Num(n) if *n >= 1.0 => (*n as usize).saturating_sub(1),
-                        Value::Int(i) => (i.to_f64() as usize).saturating_sub(1),
-                        _ => axis,
-                    };
-                    axis_explicit = true;
-                } else if let Some(input_idx) = plan.inputs.iter().position(|v| *v == dim_vid) {
-                    if let Some(cv) = plan.constants.get(&input_idx) {
-                        axis = match cv {
-                            Value::Num(n) if *n >= 1.0 => (*n as usize).saturating_sub(1),
-                            Value::Int(i) => (i.to_f64() as usize).saturating_sub(1),
-                            _ => axis,
-                        };
-                        axis_explicit = true;
-                    }
-                }
-            } else {
-                // Legacy fallback: inspect any constant mapped to the second logical position
-                if let Some(dim_const) = plan.constants.get(&1) {
-                    axis = match dim_const {
-                        Value::Num(n) if *n >= 1.0 => (*n as usize).saturating_sub(1),
-                        Value::Int(i) => (i.to_f64() as usize).saturating_sub(1),
-                        _ => axis,
-                    };
-                    axis_explicit = true;
-                }
-            }
-        }
-        let (reduce_len, num_slices) = {
-            // Try to get the data tensor's shape via the reduction builtin op input id
-            let mut rows_cols: Option<(usize, usize)> = None;
-            // Prefer shape from fusion plan reduction_data if fully known in graph
-            if let Some(shape) = plan.reduction_data_shape(graph) {
-                if shape.len() >= 2 {
-                    rows_cols = Some((shape[0].max(1), shape[1].max(1)));
-                } else if shape.len() == 1 {
-                    rows_cols = Some((shape[0].max(1), 1));
-                }
-            }
-            // Early fallback: inspect runtime variable values for declared plan inputs
-            if rows_cols.is_none() {
-                for &vid in &plan.inputs {
-                    if let Some(binding) = graph.var_binding(vid) {
-                        let value_opt = match binding.kind {
-                            VarKind::Global => vars.get(binding.index).cloned(),
-                            VarKind::Local => {
-                                if let Some(frame) = context.call_stack.last() {
-                                    let absolute = frame.locals_start + binding.index;
-                                    context.locals.get(absolute).cloned()
-                                } else {
-                                    vars.get(binding.index).cloned()
-                                }
-                            }
-                        };
-                        if let Some(value) = value_opt {
-                            match value {
-                                Value::GpuTensor(h) => {
-                                    rows_cols = Some((
-                                        h.shape.first().copied().unwrap_or(1).max(1),
-                                        h.shape.get(1).copied().unwrap_or(1).max(1),
-                                    ));
-                                    break;
-                                }
-                                Value::Tensor(t) => {
-                                    rows_cols = Some((
-                                        t.shape.first().copied().unwrap_or(1).max(1),
-                                        t.shape.get(1).copied().unwrap_or(1).max(1),
-                                    ));
-                                    break;
-                                }
-                                _ => {}
-                            }
-                        }
-                    }
-                }
-            }
-            // Prefer immediately-consumed stack values (common for reductions over producer results)
-            for v in consumed_inputs.iter().filter_map(|v| v.as_ref()) {
-                match v {
-                    Value::GpuTensor(h) => {
-                        rows_cols = Some((
-                            h.shape.first().copied().unwrap_or(1).max(1),
-                            h.shape.get(1).copied().unwrap_or(1).max(1),
-                        ));
-                        break;
-                    }
-                    Value::Tensor(t) => {
-                        rows_cols = Some((
-                            t.shape.first().copied().unwrap_or(1).max(1),
-                            t.shape.get(1).copied().unwrap_or(1).max(1),
-                        ));
-                        break;
-                    }
-                    _ => {}
-                }
-            }
-            let data_value_id: Option<runmat_accelerate::graph::ValueId> = plan.reduction_data;
-
-            if let Some(data_id) = data_value_id {
-                // Map data_id to plan input index if external
-                if let Some(input_index) = plan.inputs.iter().position(|vid| *vid == data_id) {
-                    if let Some(val) = consumed_inputs.get(input_index).and_then(|v| v.as_ref()) {
-                        match val {
-                            Value::GpuTensor(h) => {
-                                let r = h.shape.first().copied().unwrap_or(1).max(1);
-                                let c = h.shape.get(1).copied().unwrap_or(1).max(1);
-                                rows_cols = Some((r, c));
-                            }
-                            Value::Tensor(t) => {
-                                let r = t.shape.first().copied().unwrap_or(1).max(1);
-                                let c = t.shape.get(1).copied().unwrap_or(1).max(1);
-                                rows_cols = Some((r, c));
-                            }
-                            _ => {}
-                        }
-                    }
-                    // Otherwise, it was a variable/constant; use request.inputs
-                    if rows_cols.is_none() {
-                        if let Some(val) = request.inputs.get(input_index) {
-                            match val {
-                                Value::GpuTensor(h) => {
-                                    let r = h.shape.first().copied().unwrap_or(1).max(1);
-                                    let c = h.shape.get(1).copied().unwrap_or(1).max(1);
-                                    rows_cols = Some((r, c));
-                                }
-                                Value::Tensor(t) => {
-                                    let r = t.shape.first().copied().unwrap_or(1).max(1);
-                                    let c = t.shape.get(1).copied().unwrap_or(1).max(1);
-                                    rows_cols = Some((r, c));
-                                }
-                                _ => {}
-                            }
-                        }
-                    }
-                }
-                if rows_cols.is_none() {
-                    if let Some(info) = graph.value(data_id) {
-                        // Try direct variable lookup to get runtime value shape
-                        if let ValueOrigin::Variable { kind, index } = &info.origin {
-                            let val = match kind {
-                                VarKind::Global => vars.get(*index).cloned(),
-                                VarKind::Local => {
-                                    if let Some(frame) = context.call_stack.last() {
-                                        let absolute = frame.locals_start + index;
-                                        context.locals.get(absolute).cloned()
-                                    } else {
-                                        vars.get(*index).cloned()
-                                    }
-                                }
-                            };
-                            if let Some(v) = val {
-                                match v {
-                                    Value::GpuTensor(h) => {
-                                        rows_cols = Some((
-                                            h.shape.first().copied().unwrap_or(1).max(1),
-                                            h.shape.get(1).copied().unwrap_or(1).max(1),
-                                        ));
-                                    }
-                                    Value::Tensor(t) => {
-                                        rows_cols = Some((
-                                            t.shape.first().copied().unwrap_or(1).max(1),
-                                            t.shape.get(1).copied().unwrap_or(1).max(1),
-                                        ));
-                                    }
-                                    _ => {}
-                                }
-                            }
-                        }
-                        if rows_cols.is_none() {
-                            if let ShapeInfo::Tensor(dims) = &info.shape {
-                                if !dims.is_empty() {
-                                    let r = dims.first().and_then(|d| *d).unwrap_or(1);
-                                    let c = dims.get(1).and_then(|d| *d).unwrap_or(1);
-                                    rows_cols = Some((r.max(1), c.max(1)));
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            // Fallback: any tensor input
-            if rows_cols.is_none() {
-                for v in consumed_inputs.iter().filter_map(|v| v.as_ref()) {
-                    match v {
-                        Value::GpuTensor(h) => {
-                            rows_cols = Some((
-                                h.shape.first().copied().unwrap_or(1).max(1),
-                                h.shape.get(1).copied().unwrap_or(1).max(1),
-                            ));
-                            break;
-                        }
-                        Value::Tensor(t) => {
-                            rows_cols = Some((
-                                t.shape.first().copied().unwrap_or(1).max(1),
-                                t.shape.get(1).copied().unwrap_or(1).max(1),
-                            ));
-                            break;
-                        }
-                        _ => {}
-                    }
-                }
-                if rows_cols.is_none() {
-                    for v in &request.inputs {
-                        match v {
-                            Value::GpuTensor(h) => {
-                                rows_cols = Some((
-                                    h.shape.first().copied().unwrap_or(1).max(1),
-                                    h.shape.get(1).copied().unwrap_or(1).max(1),
-                                ));
-                                break;
-                            }
-                            Value::Tensor(t) => {
-                                rows_cols = Some((
-                                    t.shape.first().copied().unwrap_or(1).max(1),
-                                    t.shape.get(1).copied().unwrap_or(1).max(1),
-                                ));
-                                break;
-                            }
-                            _ => {}
-                        }
-                    }
-                }
-            }
-            // Final fallback: group-level static shape if available
-            if rows_cols.is_none() {
-                if let ShapeInfo::Tensor(dims) = &plan.group.shape {
-                    if !dims.is_empty() {
-                        let r = dims.first().and_then(|d| *d).unwrap_or(1);
-                        let c = dims.get(1).and_then(|d| *d).unwrap_or(1);
-                        rows_cols = Some((r.max(1), c.max(1)));
-                    }
-                }
-            }
-
-            let (r, c) = rows_cols.unwrap_or((1, 1));
-            if reduce_all {
-                let mut total_elems: Option<usize> = None;
-                let mut total_from_operand = false;
-                // Prefer fully-known graph shape for the reduction operand
-                if let Some(shape) = plan.reduction_data_shape(graph) {
-                    let prod = shape.into_iter().fold(1usize, |acc, dim| {
-                        let d = dim.max(1);
-                        acc.saturating_mul(d)
-                    });
-                    total_from_operand = true;
-                    total_elems = Some(prod.max(1));
-                }
-                // Fall back to runtime tensor shapes (consumed stack values first, then inputs)
-                if total_elems.is_none() {
-                    let inspect_value = |value: &Value| -> Option<usize> {
-                        match value {
-                            Value::GpuTensor(handle) => {
-                                if handle.shape.is_empty() {
-                                    Some(1)
-                                } else {
-                                    Some(
-                                        handle
-                                            .shape
-                                            .iter()
-                                            .copied()
-                                            .map(|d| d.max(1))
-                                            .fold(1usize, |acc, dim| acc.saturating_mul(dim)),
-                                    )
-                                }
-                            }
-                            Value::Tensor(tensor) => {
-                                if tensor.shape.is_empty() {
-                                    Some(1)
-                                } else {
-                                    Some(
-                                        tensor
-                                            .shape
-                                            .iter()
-                                            .copied()
-                                            .map(|d| d.max(1))
-                                            .fold(1usize, |acc, dim| acc.saturating_mul(dim)),
-                                    )
-                                }
-                            }
-                            _ => None,
-                        }
-                    };
-                    for value in consumed_inputs.iter().filter_map(|v| v.as_ref()) {
-                        if let Some(prod) = inspect_value(value) {
-                            total_from_operand = true;
-                            total_elems = Some(prod.max(1));
-                            break;
-                        }
-                    }
-                    if total_elems.is_none() {
-                        for value in &request.inputs {
-                            if let Some(prod) = inspect_value(value) {
-                                total_from_operand = true;
-                                total_elems = Some(prod.max(1));
-                                break;
-                            }
-                        }
-                    }
-                }
-                // Final fallback: use group-level element count or the 2-D heuristic
-                if total_elems.is_none() {
-                    if let Some(ec) = plan.element_count() {
-                        total_elems = Some(ec.max(1));
-                    }
-                }
-                if total_elems.is_none() || !total_from_operand {
-                    if interp_engine::fusion_debug_enabled() {
-                        log::debug!(
-                            "fusion reduction (all): operand extent unknown (source: {:?}); falling back to provider path",
-                            if total_from_operand { "runtime" } else { "output_shape" }
-                        );
-                    }
-                    return Err(mex(
-                        "FusionReductionExtentUnknown",
-                        "fusion: reduction all extent unknown",
-                    ));
-                }
-                let total = total_elems.unwrap();
-                if interp_engine::fusion_debug_enabled() {
-                    log::debug!(
-                        "fusion reduction (all): total_elems={} fallback_rows={} fallback_cols={}",
-                        total,
-                        r,
-                        c
-                    );
-                }
-                (total, 1usize)
-            } else {
-                if !axis_explicit {
-                    axis = if r == 1 && c > 1 {
-                        1
-                    } else if r > 1 {
-                        0
-                    } else {
-                        axis
-                    };
-                }
-                if interp_engine::fusion_debug_enabled() {
-                    if r == 1 && c == 1 {
-                        log::debug!(
-                    "fusion reduction: unresolved shape (defaulted to 1x1); axis={}, constants={:?}",
-                    axis, plan.constants
-                );
-                    } else {
-                        log::debug!(
-                    "fusion reduction: resolved shape rows={} cols={} axis={} constants={:?}",
-                    r,
-                    c,
-                    axis,
-                    plan.constants
-                );
-                    }
-                }
-                if axis == 0 {
-                    (r, c)
-                } else {
-                    (c, r)
-                }
-            }
-        };
-        if interp_engine::fusion_debug_enabled() {
-            log::debug!(
-                "fusion reduction: axis={} reduce_len={} num_slices={} constants={:?}",
-                axis,
-                reduce_len,
-                num_slices,
-                plan.constants
-            );
-        }
-        if log::log_enabled!(log::Level::Debug) && interp_engine::fusion_debug_enabled() {
-            let _rt_inputs: Vec<String> = request
-                .inputs
-                .iter()
-                .enumerate()
-                .map(|(i, v)| accel_fusion::summarize_value(i, v))
-                .collect();
-            let _plan_inputs: Vec<String> = plan
-                .inputs
-                .iter()
-                .map(|vid| {
-                    if let Some(info) = graph.value(*vid) {
-                        format!(
-                            "vid={} origin={:?} shape={:?}",
-                            vid, info.origin, info.shape
-                        )
-                    } else {
-                        format!("vid={} origin=<missing>", vid)
-                    }
-                })
-                .collect();
-            // Summarize inputs once before execution (omit plan inputs to reduce noise)
-            log::debug!("reduction inputs: [{}]", _rt_inputs.join(", "));
-        }
-        // If shape derivation failed (1x1) but inputs/consumed suggest a larger tensor, skip fusion
-        let looks_wrong = reduce_len == 1 && num_slices == 1 && {
-            let mut big = false;
-            let mut check_val = |v: &Value| match v {
-                Value::GpuTensor(h) => {
-                    let prod = h.shape.iter().copied().product::<usize>();
-                    if prod > 1 {
-                        big = true;
-                    }
-                }
-                Value::Tensor(t) => {
-                    let prod = t.shape.iter().copied().product::<usize>();
-                    if prod > 1 {
-                        big = true;
-                    }
-                }
-                _ => {}
-            };
-            for v in consumed_inputs.iter().filter_map(|v| v.as_ref()) {
-                check_val(v);
-            }
-            for v in &request.inputs {
-                check_val(v);
-            }
-            big
-        };
-        if looks_wrong {
-            log::debug!(
-                "fusion reduction: skipping fusion due to unresolved shape; falling back to provider path"
-            );
-            return Err(mex(
-                "FusionReductionShapeUnresolved",
-                "fusion: reduction shape unresolved",
-            ));
-        }
-
-        // Optional escape hatch: disable fused reductions to force provider path
-        if std::env::var("RUNMAT_DISABLE_FUSED_REDUCTION")
-            .ok()
-            .as_deref()
-            == Some("1")
-        {
-            return Err(mex(
-                "FusionReductionDisabled",
-                "fusion: fused reductions disabled",
-            ));
-        }
-        let workgroup_size = 256u32;
-        if log::log_enabled!(log::Level::Debug) && interp_engine::fusion_debug_enabled() {
-            let _rt_inputs: Vec<String> = request
-                .inputs
-                .iter()
-                .enumerate()
-                .map(|(i, v)| accel_fusion::summarize_value(i, v))
-                .collect();
-            let _plan_inputs: Vec<String> = plan
-                .inputs
-                .iter()
-                .map(|vid| {
-                    if let Some(info) = graph.value(*vid) {
-                        format!(
-                            "vid={} origin={:?} shape={:?}",
-                            vid, info.origin, info.shape
-                        )
-                    } else {
-                        format!("vid={} origin=<missing>", vid)
-                    }
-                })
-                .collect();
-            log::debug!(
-                "reduction axis={} reduce_len={} num_slices={}",
-                axis,
-                reduce_len,
-                num_slices
-            );
-        }
-        match execute_reduction(request, reduce_len, num_slices, workgroup_size) {
-            Ok(result) => {
-                stack_guard.commit();
-                Ok(result)
-            }
-            Err(err) => Err(mex("FusionExecutionFailed", &err.to_string())),
-        }
+        accel_fusion::execute_fusion_reduction(
+            plan,
+            graph,
+            request,
+            &consumed_inputs,
+            stack_guard,
+            vars,
+            context,
+        )
     } else {
         accel_fusion::execute_fusion_special_kind(plan.group.kind.clone(), &plan.inputs, request, stack_guard).await
     }
