+++ crates/runmat-vm/src/accel/fusion.rs
@@ -8,10 +8,11 @@ use runmat_accelerate::fusion::FusionStoreMaterialization;
 use runmat_accelerate::fusion_exec::{
     execute_centered_gram, execute_elementwise, execute_explained_variance,
     execute_image_normalize, execute_matmul_epilogue, execute_power_step_normalize,
+    execute_reduction,
     FusionExecutionRequest,
 };
 use runmat_accelerate::InstrSpan;
-use runmat_accelerate::{FusionKind, ValueOrigin, VarKind};
+use runmat_accelerate::{FusionKind, ShapeInfo, ValueOrigin, VarKind, value_is_all_keyword};
 use runmat_builtins::Value;
 use runmat_runtime::RuntimeError;
 use std::collections::HashMap;
@@ -495,5 +496,335 @@ pub async fn execute_fusion_special_kind(
             Err(err) => Err(mex("FusionExecutionFailed", &err.to_string())),
         },
         _ => Err(mex("FusionUnsupportedKind", "fusion: unsupported fusion kind")),
+    }
+}
+
+pub struct ReductionGeometry {
+    pub axis: usize,
+    pub reduce_len: usize,
+    pub num_slices: usize,
+}
+
+pub fn resolve_reduction_geometry(
+    plan: &runmat_accelerate::FusionGroupPlan,
+    graph: &runmat_accelerate::AccelGraph,
+    request: &FusionExecutionRequest<'_>,
+    consumed_inputs: &[Option<Value>],
+    vars: &Vec<Value>,
+    context: &ExecutionContext,
+) -> Result<ReductionGeometry, RuntimeError> {
+    fn detect_reduce_all(
+        plan: &runmat_accelerate::FusionGroupPlan,
+        graph: &runmat_accelerate::AccelGraph,
+    ) -> bool {
+        let mut reduce_all = matches!(plan.reduction_axes, Some(runmat_accelerate::ReductionAxes::All));
+        let has_all = reduce_all
+            || plan.constants.values().any(value_is_all_keyword)
+            || plan.const_values.values().any(value_is_all_keyword);
+        if has_all {
+            return true;
+        }
+        for node_id in &plan.group.nodes {
+            if let Some(node) = graph.node(*node_id) {
+                if let runmat_accelerate::graph::AccelNodeLabel::Builtin { name } = &node.label {
+                    if name.eq_ignore_ascii_case("mean") {
+                        for input_vid in &node.inputs {
+                            if let Some(info) = graph.value(*input_vid) {
+                                if let Some(constant) = &info.constant {
+                                    if value_is_all_keyword(constant) {
+                                        reduce_all = true;
+                                        break;
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            if reduce_all { break; }
+        }
+        reduce_all
+    }
+
+    fn resolve_reduction_axis(plan: &runmat_accelerate::FusionGroupPlan) -> (usize, bool) {
+        let mut axis = 0usize;
+        let mut axis_explicit = false;
+        if let Some(runmat_accelerate::ReductionAxes::Explicit(dims)) = &plan.reduction_axes {
+            if let Some(first) = dims.first().copied() {
+                axis = first.saturating_sub(1);
+                axis_explicit = true;
+            }
+        }
+        if let Some(dim_vid) = plan.reduction_dim {
+            if let Some(cv) = plan.const_values.get(&dim_vid) {
+                axis = match cv {
+                    Value::Num(n) if *n >= 1.0 => (*n as usize).saturating_sub(1),
+                    Value::Int(i) => (i.to_f64() as usize).saturating_sub(1),
+                    _ => axis,
+                };
+                axis_explicit = true;
+            } else if let Some(input_idx) = plan.inputs.iter().position(|v| *v == dim_vid) {
+                if let Some(cv) = plan.constants.get(&input_idx) {
+                    axis = match cv {
+                        Value::Num(n) if *n >= 1.0 => (*n as usize).saturating_sub(1),
+                        Value::Int(i) => (i.to_f64() as usize).saturating_sub(1),
+                        _ => axis,
+                    };
+                    axis_explicit = true;
+                }
+            }
+        } else if let Some(dim_const) = plan.constants.get(&1) {
+            axis = match dim_const {
+                Value::Num(n) if *n >= 1.0 => (*n as usize).saturating_sub(1),
+                Value::Int(i) => (i.to_f64() as usize).saturating_sub(1),
+                _ => axis,
+            };
+            axis_explicit = true;
+        }
+        (axis, axis_explicit)
+    }
+
+    fn derive_rows_cols(
+        plan: &runmat_accelerate::FusionGroupPlan,
+        graph: &runmat_accelerate::AccelGraph,
+        request: &FusionExecutionRequest<'_>,
+        consumed_inputs: &[Option<Value>],
+        vars: &Vec<Value>,
+        context: &ExecutionContext,
+    ) -> Option<(usize, usize)> {
+        let shape_of = |value: &Value| -> Option<(usize, usize)> {
+            match value {
+                Value::GpuTensor(h) => Some((
+                    h.shape.first().copied().unwrap_or(1).max(1),
+                    h.shape.get(1).copied().unwrap_or(1).max(1),
+                )),
+                Value::Tensor(t) => Some((
+                    t.shape.first().copied().unwrap_or(1).max(1),
+                    t.shape.get(1).copied().unwrap_or(1).max(1),
+                )),
+                _ => None,
+            }
+        };
+
+        if let Some(shape) = plan.reduction_data_shape(graph) {
+            if shape.len() >= 2 {
+                return Some((shape[0].max(1), shape[1].max(1)));
+            }
+            if shape.len() == 1 {
+                return Some((shape[0].max(1), 1));
+            }
+        }
+
+        for &vid in &plan.inputs {
+            if let Some(binding) = graph.var_binding(vid) {
+                let value_opt = match binding.kind {
+                    VarKind::Global => vars.get(binding.index).cloned(),
+                    VarKind::Local => {
+                        if let Some(frame) = context.call_stack.last() {
+                            let absolute = frame.locals_start + binding.index;
+                            context.locals.get(absolute).cloned()
+                        } else {
+                            vars.get(binding.index).cloned()
+                        }
+                    }
+                };
+                if let Some(value) = value_opt {
+                    if let Some(shape) = shape_of(&value) {
+                        return Some(shape);
+                    }
+                }
+            }
+        }
+
+        for v in consumed_inputs.iter().filter_map(|v| v.as_ref()) {
+            if let Some(shape) = shape_of(v) {
+                return Some(shape);
+            }
+        }
+
+        if let Some(data_id) = plan.reduction_data {
+            if let Some(input_index) = plan.inputs.iter().position(|vid| *vid == data_id) {
+                if let Some(val) = consumed_inputs.get(input_index).and_then(|v| v.as_ref()) {
+                    if let Some(shape) = shape_of(val) {
+                        return Some(shape);
+                    }
+                }
+                if let Some(val) = request.inputs.get(input_index) {
+                    if let Some(shape) = shape_of(val) {
+                        return Some(shape);
+                    }
+                }
+            }
+            if let Some(info) = graph.value(data_id) {
+                if let ValueOrigin::Variable { kind, index } = &info.origin {
+                    let val = match kind {
+                        VarKind::Global => vars.get(*index).cloned(),
+                        VarKind::Local => {
+                            if let Some(frame) = context.call_stack.last() {
+                                let absolute = frame.locals_start + index;
+                                context.locals.get(absolute).cloned()
+                            } else {
+                                vars.get(*index).cloned()
+                            }
+                        }
+                    };
+                    if let Some(v) = val {
+                        if let Some(shape) = shape_of(&v) {
+                            return Some(shape);
+                        }
+                    }
+                }
+                if let ShapeInfo::Tensor(dims) = &info.shape {
+                    if !dims.is_empty() {
+                        let r = dims.first().and_then(|d| *d).unwrap_or(1);
+                        let c = dims.get(1).and_then(|d| *d).unwrap_or(1);
+                        return Some((r.max(1), c.max(1)));
+                    }
+                }
+            }
+        }
+
+        for v in &request.inputs {
+            if let Some(shape) = shape_of(v) {
+                return Some(shape);
+            }
+        }
+
+        if let ShapeInfo::Tensor(dims) = &plan.group.shape {
+            if !dims.is_empty() {
+                let r = dims.first().and_then(|d| *d).unwrap_or(1);
+                let c = dims.get(1).and_then(|d| *d).unwrap_or(1);
+                return Some((r.max(1), c.max(1)));
+            }
+        }
+        None
+    }
+
+    if log::log_enabled!(log::Level::Debug) {
+        let meta: Vec<String> = plan.inputs.iter().map(|vid| {
+            if let Some(info) = graph.value(*vid) {
+                format!("vid={} origin={:?} shape={:?}", vid, info.origin, info.shape)
+            } else {
+                format!("vid={} origin=<missing>", vid)
+            }
+        }).collect();
+        log::debug!("reduction gather meta: [{}]", meta.join(", "));
+    }
+
+    let reduce_all = detect_reduce_all(plan, graph);
+    let (mut axis, axis_explicit) = if reduce_all { (0usize, false) } else { resolve_reduction_axis(plan) };
+    if reduce_all && interp_engine::fusion_debug_enabled() {
+        log::debug!(
+            "fusion reduction (all) meta: data_vid={:?} inputs={:?} stack_pattern={:?}",
+            plan.reduction_data,
+            plan.inputs,
+            plan.stack_pattern
+        );
+    }
+
+    let (r, c) = derive_rows_cols(plan, graph, request, consumed_inputs, vars, context).unwrap_or((1, 1));
+    let (reduce_len, num_slices) = if reduce_all {
+        let total_from_runtime = consumed_inputs
+            .iter()
+            .filter_map(|v| v.as_ref())
+            .chain(request.inputs.iter())
+            .find_map(|value| match value {
+                Value::GpuTensor(handle) => Some(if handle.shape.is_empty() { 1 } else { handle.shape.iter().copied().map(|d| d.max(1)).product::<usize>() }),
+                Value::Tensor(tensor) => Some(if tensor.shape.is_empty() { 1 } else { tensor.shape.iter().copied().map(|d| d.max(1)).product::<usize>() }),
+                _ => None,
+            });
+        let total = plan.reduction_data_shape(graph)
+            .map(|shape| shape.into_iter().map(|d| d.max(1)).product::<usize>())
+            .or(total_from_runtime)
+            .or_else(|| plan.element_count())
+            .filter(|v| *v > 0)
+            .ok_or_else(|| mex("FusionReductionExtentUnknown", "fusion: reduction all extent unknown"))?;
+        if interp_engine::fusion_debug_enabled() {
+            log::debug!(
+                "fusion reduction (all): total_elems={} fallback_rows={} fallback_cols={}",
+                total,
+                r,
+                c
+            );
+        }
+        (total, 1usize)
+    } else {
+        if !axis_explicit {
+            axis = if r == 1 && c > 1 { 1 } else if r > 1 { 0 } else { axis };
+        }
+        if interp_engine::fusion_debug_enabled() {
+            if r == 1 && c == 1 {
+                log::debug!(
+                    "fusion reduction: unresolved shape (defaulted to 1x1); axis={}, constants={:?}",
+                    axis,
+                    plan.constants
+                );
+            } else {
+                log::debug!(
+                    "fusion reduction: resolved shape rows={} cols={} axis={} constants={:?}",
+                    r,
+                    c,
+                    axis,
+                    plan.constants
+                );
+            }
+        }
+        if axis == 0 { (r, c) } else { (c, r) }
+    };
+
+    if interp_engine::fusion_debug_enabled() {
+        log::debug!(
+            "fusion reduction: axis={} reduce_len={} num_slices={} constants={:?}",
+            axis,
+            reduce_len,
+            num_slices,
+            plan.constants
+        );
+    }
+
+    let looks_wrong = reduce_len == 1 && num_slices == 1 && {
+        let mut big = false;
+        let mut check_val = |v: &Value| match v {
+            Value::GpuTensor(h) => {
+                let prod = h.shape.iter().copied().product::<usize>();
+                if prod > 1 { big = true; }
+            }
+            Value::Tensor(t) => {
+                let prod = t.shape.iter().copied().product::<usize>();
+                if prod > 1 { big = true; }
+            }
+            _ => {}
+        };
+        for v in consumed_inputs.iter().filter_map(|v| v.as_ref()) { check_val(v); }
+        for v in &request.inputs { check_val(v); }
+        big
+    };
+    if looks_wrong {
+        log::debug!("fusion reduction: skipping fusion due to unresolved shape; falling back to provider path");
+        return Err(mex("FusionReductionShapeUnresolved", "fusion: reduction shape unresolved"));
+    }
+    if std::env::var("RUNMAT_DISABLE_FUSED_REDUCTION").ok().as_deref() == Some("1") {
+        return Err(mex("FusionReductionDisabled", "fusion: fused reductions disabled"));
+    }
+
+    Ok(ReductionGeometry { axis, reduce_len, num_slices })
+}
+
+pub fn execute_fusion_reduction(
+    plan: &runmat_accelerate::FusionGroupPlan,
+    graph: &runmat_accelerate::AccelGraph,
+    request: FusionExecutionRequest<'_>,
+    consumed_inputs: &[Option<Value>],
+    stack_guard: StackSliceGuard<'_>,
+    vars: &Vec<Value>,
+    context: &ExecutionContext,
+) -> Result<Value, RuntimeError> {
+    let geom = resolve_reduction_geometry(plan, graph, &request, consumed_inputs, vars, context)?;
+    match execute_reduction(request, geom.reduce_len, geom.num_slices, 256u32) {
+        Ok(result) => {
+            stack_guard.commit();
+            Ok(result)
+        }
+        Err(err) => Err(mex("FusionExecutionFailed", &err.to_string())),
     }
 }
